imagecaption/requirements.txt
opencv-python==4.7.0.72
requests==2.30.0
kafka-python==2.0.2